integrate protobuf
Former-commit-id: 73ac72e859f8a4fc8b14a4bf427eea9694f7e932

+++ src/node/Node.h
@@ -19,7 +19,7 @@ private:
     int listeningSocket;
 
     std::unordered_set<std::string> nodeAddresses = {"10.0.1.4", "10.0.1.5"};//, "10.0.1.5", "10.0.1.6", "10.0.1.7", "10.0.1.8"};
-    size_t numNodes;
+    int numNodes;
     std::unordered_map<std::string, int> readSockets;
     std::unordered_map<std::string, int> writeSockets;
